youtube_to_list/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
google-generativeai
youtube-transcript-api
jinja2